+++ insights-ui/package.json
@@ -16,7 +16,7 @@
     "node": ">=20.0.0"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.758.0",
+    "@aws-sdk/client-s3": "3.363.0",
     "ajv": "^8.17.1",
     "chart.js": "^4.4.7",
     "gray-matter": "^4.0.3",
