+++ insights-ui/src/app/api/actions/criteria/copy-ai-criteria/route.ts
@@ -1,7 +1,8 @@
 // app/api/public-equity/copy-ai-criteria/route.ts
+import { CriteriaLookupItem } from '@/types/public-equity/ticker-report';
+import { CreateCriteriaRequest } from '@/types/public-equity/ticker-request-response';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
-import { CreateCriteriaRequest, CriteriaLookupItem } from '@/lib/publicEquity';
 import {
   getCriteriaLookupList,
   getMatchingCriteriaLookupItem,
