+++ insights-ui/src/app/api/actions/tickers/[ticker]/trigger-criteria-matching/route.ts
@@ -1,13 +1,5 @@
-import {
-  CriterionMatchesOfLatest10Q,
-  getCriteria,
-  getTickerReport,
-  initializeNewTickerReport,
-  saveTickerReport,
-  TickerReport,
-  triggerCriteriaMatching,
-} from '@/lib/publicEquity';
-import { ProcessingStatus } from '@/types/public-equity/ticker-report';
+import { getTickerReport, initializeNewTickerReport, saveTickerReport, triggerCriteriaMatching } from '@/lib/publicEquity';
+import { ProcessingStatus, TickerReport } from '@/types/public-equity/ticker-report';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 
 // app/api/public-equity/single-criterion-report/route.ts
@@ -15,15 +7,15 @@ import { NextRequest } from 'next/server';
 
 const triggerCriteriaMatchingForTicker = async (req: NextRequest, { params }: { params: Promise<{ tickerKey: string }> }): Promise<TickerReport> => {
   const { tickerKey } = await params;
-  const tickerReport = await getTickerReport(body.ticker);
+  const tickerReport = await getTickerReport(tickerKey);
   await initializeNewTickerReport(tickerKey, tickerReport.selectedSector.id, tickerReport.selectedIndustryGroup.id);
   await triggerCriteriaMatching(tickerKey, false);
-  const criteria = await getCriteria(tickerReport.selectedSector.name, tickerReport.selectedIndustryGroup.name);
-  const criteriaMatches: CriterionMatchesOfLatest10Q[] = criteria.criteria.map((criterion) => ({
-    criterionKey: criterion.key,
-    status: ProcessingStatus.InProgress,
-  }));
-  const updatedReport = { ...tickerReport, criteriaMatchesOfLatest10Q: criteriaMatches };
+  const updatedReport: TickerReport = {
+    ...tickerReport,
+    criteriaMatchesOfLatest10Q: {
+      status: ProcessingStatus.InProgress,
+    },
+  };
   await saveTickerReport(tickerKey, updatedReport);
 
   return updatedReport;
