+++ insights-ui/src/app/api/ai-criteria/route.ts
@@ -1,12 +1,6 @@
-import {
-  CreateCriteriaRequest,
-  generateAiCriteria,
-  getCriteriaLookupList,
-  getMatchingCriteriaLookupItem,
-  IndustryGroupCriteria,
-  updateCriteriaLookupList,
-  uploadAiCriteriaToS3,
-} from '@/lib/publicEquity';
+import { generateAiCriteria, getCriteriaLookupList, getMatchingCriteriaLookupItem, updateCriteriaLookupList, uploadAiCriteriaToS3 } from '@/lib/publicEquity';
+import { IndustryGroupCriteria } from '@/types/public-equity/ticker-report';
+import { CreateCriteriaRequest } from '@/types/public-equity/ticker-request-response';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { NextRequest } from 'next/server';
