+++ insights-ui/src/app/api/tickers/[tickerKey]/route.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@/prisma';
-import { TickerUpsertRequest } from '@/types/public-equity/ticker';
+import { TickerUpsertRequest } from '@/types/public-equity/ticker-request-response';
 import { withErrorHandlingV2 } from '@dodao/web-core/api/helpers/middlewares/withErrorHandling';
 import { Ticker } from '@prisma/client';
 import { NextRequest } from 'next/server';
