+++ insights-ui/src/app/public-equities/industry-group-criteria/[sectorSlug]/[industryGroupSlug]/create/page.tsx
@@ -1,6 +1,7 @@
 import CriteriaTable from '@/components/criteria/CriteriaTable';
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
-import { IndustryGroupCriteria } from '@/types/criteria/criteria';
+
+import { IndustryGroupCriteria } from '@/types/public-equity/ticker-report';
 import { BreadcrumbsOjbect } from '@dodao/web-core/components/core/breadcrumbs/BreadcrumbsWithChevrons'; // Adjust the import path as needed
 import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
