+++ insights-ui/src/app/public-equities/tickers/[tickerKey]/page.tsx
@@ -1,7 +1,13 @@
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
 import RadarChart from '@/components/ui/RadarChart';
-import { IndustryGroupCriteria } from '@/types/criteria/criteria';
-import { CriteriaEvaluation, PerformanceChecklistItem, SpiderGraphForTicker, SpiderGraphPie, TickerReport } from '@/types/public-equity/ticker-report';
+import {
+  CriteriaEvaluation,
+  IndustryGroupCriteria,
+  PerformanceChecklistItem,
+  SpiderGraphForTicker,
+  SpiderGraphPie,
+  TickerReport,
+} from '@/types/public-equity/ticker-report';
 import { getReportName } from '@/util/report-utils';
 import { BreadcrumbsOjbect } from '@dodao/web-core/components/core/breadcrumbs/BreadcrumbsWithChevrons';
 import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
