+++ insights-ui/src/types/public-equity/ticker-report.ts
@@ -14,23 +14,6 @@ export interface Sector {
   name: string;
 }
 
-export interface Metric {
-  metricKey: string;
-  value: number;
-  calculationExplanation: string;
-}
-
-export interface ImportantMetrics {
-  status: ProcessingStatus;
-  metrics: Metric[];
-}
-
-export interface CriterionReport {
-  reportKey?: string;
-  status?: ProcessingStatus;
-  outputFileUrl?: string | null;
-}
-
 export interface PerformanceChecklistItem {
   checklistItem: string;
   oneLinerExplanation: string;
@@ -38,13 +21,6 @@ export interface PerformanceChecklistItem {
   detailedExplanation: string;
   evaluationLogic: string;
   score: number;
-}
-
-export interface CriteriaEvaluation {
-  criterionKey: string;
-  importantMetrics?: ImportantMetrics;
-  reports?: CriterionReport[];
-  performanceChecklist?: PerformanceChecklistItem[];
 }
 
 export interface SecFilingAttachment {
@@ -62,23 +38,9 @@ export interface CriterionMatch {
 }
 
 export interface CriterionMatchesOfLatest10Q {
-  criterionMatches: CriterionMatch[];
+  criterionMatches?: CriterionMatch[];
   status: ProcessingStatus;
   failureReason?: string;
-}
-
-export interface TickerReport {
-  ticker: string;
-  selectedIndustryGroup: IndustryGroup;
-  selectedSector: Sector;
-  evaluationsOfLatest10Q?: CriteriaEvaluation[];
-  criteriaMatchesOfLatest10Q?: CriterionMatchesOfLatest10Q;
-}
-
-export interface CreateAllReportsRequest {
-  ticker: string;
-  industryGroupId: number;
-  sectorId: number;
 }
 
 export interface CreateSingleReportsRequest {
@@ -106,6 +68,103 @@ export interface RegenerateSingleCriterionReportsRequest {
   ticker: string;
   criterionKey: string;
 }
-export interface RegenerateAllCriteriaReportsRequest {
+
+export interface MetricValueItem {
+  metricKey: string;
+  value: string;
+  calculationExplanation: string;
+}
+
+export interface ImportantMetricsValue {
+  status: ProcessingStatus;
+  metrics?: MetricValueItem[];
+}
+
+export interface CriterionReportValueItem {
+  reportKey: string;
+  status: ProcessingStatus;
+  outputFileUrl?: string;
+}
+
+export interface PerformanceChecklistItem {
+  checklistItem: string;
+  oneLinerExplanation: string;
+  informationUsed: string;
+  detailedExplanation: string;
+  evaluationLogic: string;
+  score: number;
+}
+
+export interface CriteriaEvaluation {
+  criterionKey: string;
+  importantMetrics?: ImportantMetricsValue;
+  reports?: CriterionReportValueItem[];
+  performanceChecklist?: PerformanceChecklistItem[];
+}
+
+export interface SecFilingAttachment {
+  attachmentSequenceNumber: string;
+  attachmentDocumentName: string;
+  attachmentPurpose?: string;
+  attachmentUrl: string;
+  matchedPercentage: number;
+  latest10QContent: string;
+}
+
+export interface CriterionMatch {
+  criterionKey: string;
+  matchedAttachments: SecFilingAttachment[];
+  matchedContent: string;
+}
+
+export interface TickerReport {
   ticker: string;
+  selectedIndustryGroup: IndustryGroup;
+  selectedSector: Sector;
+  evaluationsOfLatest10Q?: CriteriaEvaluation[];
+  criteriaMatchesOfLatest10Q?: CriterionMatchesOfLatest10Q;
+}
+
+export interface MetricDefinitionItem {
+  key: string;
+  name: string;
+  description: string;
+  formula: string;
+}
+
+export type OutputType = 'Text' | 'BarGraph' | 'PieChart' | 'WaterfallChart';
+
+export interface CriterionReportDefinitionItem {
+  key: string;
+  name: string;
+  description: string;
+  outputType: OutputType;
+}
+
+export interface IndustryGroupCriterion {
+  key: string;
+  name: string;
+  shortDescription: string;
+  importantMetrics: MetricDefinitionItem[];
+  reports: CriterionReportDefinitionItem[];
+}
+
+export interface IndustryGroupCriteria {
+  tickers: string[];
+  selectedSector: Sector;
+  selectedIndustryGroup: IndustryGroup;
+  criteria: IndustryGroupCriterion[];
+}
+
+export interface CriteriaLookupItem {
+  sectorId: number;
+  sectorName: string;
+  industryGroupId: number;
+  industryGroupName: string;
+  aiCriteriaFileUrl?: string;
+  customCriteriaFileUrl?: string;
+}
+
+export interface CriteriaLookupList {
+  criteria: CriteriaLookupItem[];
 }
