--- insights-ui/src/types/public-equity/ticker.ts
@@ -1,13 +1,0 @@
-export interface TickerUpsertRequest {
-  tickerKey: string;
-  sectorId: number;
-  industryGroupId: number;
-  reportUrl?: string;
-}
-
-export interface TickerCreateRequest {
-  tickerKey: string;
-  sectorId: number;
-  industryGroupId: number;
-  reportUrl?: string;
-}
